Remove warnings when compiling for 64 bit targets

+++ iConsole/iConsole.m
@@ -116,7 +116,7 @@ void exceptionHandler(NSException *exception)
 	NSInteger touches = (TARGET_IPHONE_SIMULATOR ? _simulatorTouchesToShow: _deviceTouchesToShow);
 	if (touches > 0 && touches < 11)
 	{
-		text = [text stringByAppendingFormat:@"\nSwipe down with %i finger%@ to hide console", touches, (touches != 1)? @"s": @""];
+		text = [text stringByAppendingFormat:@"\nSwipe down with %li finger%@ to hide console", (long)touches, (touches != 1)? @"s": @""];
 	}
 	else if (TARGET_IPHONE_SIMULATOR ? _simulatorShakeToShow: _deviceShakeToShow)
 	{
